Generate Monte Carlo Estimates

Get a list of the output of lavaan::sem() and generate Monte Carlo estimates of model parameters.

Usage

gen_mc(
  fit_all,
  R = 100,
  ...,
  iseed = NULL,
  parallel = FALSE,
  progress = FALSE,
  ncores = max(1, parallel::detectCores(logical = FALSE) - 1),
  compute_implied_stats = FALSE
)

Arguments

fit_all

The output of fit_model() or an object of the class fit_out.

R

The number of replications to generate the Monte Carlo estimates for each fit output.

...

Optional arguments to be passed to manymome::do_mc() when generating the Monte Carlo estimates.

iseed

The seed for the random number generator. Default is NULL and the seed is not changed.

parallel

If TRUE, parallel processing will be used to generate Monte Carlo estimates for the fit outputs. Default is FALSE.

progress

If TRUE, the progress will be displayed. Default is `FALSE.

ncores

The number of CPU cores to use if parallel processing is used.

compute_implied_stats

Whether implied statistics are computed in each Monte Carlo replication. Usually not needed and so default to FALSE Required manymome 0.2.7.1 or above.

Value

A mc_list object, which is a list of the output of manymome::do_mc().

Details

It simply calls manymome::do_mc() on each output of lavaan::sem() in fit_all. The simulated estimates can then be used to test effects such as indirect effects, usually by functions from the manymome package, such as manymome::indirect_effect().

This function is used by the all-in-one function power4test(). Users usually do not call this function directly.

See also

Examples

mod <-
"m ~ x
 y ~ m + x"
es <-
c("y ~ m" = "m",
  "m ~ x" = "m",
  "y ~ x" = "n")
data_all <- sim_data(nrep = 5,
                     model = mod,
                     pop_es = es,
                     n = 100,
                     iseed = 1234)

fit_all <- fit_model(data_all)
mc_all <- gen_mc(fit_all,
                 R = 100,
                 iseed = 4567)