Rejection Rates of All Tests

Get all rejection rates of all tests stored in a power4test object.

Usage

get_rejection_rates(object, all_columns = FALSE)

Arguments

object

A power4test object.

all_columns

If TRUE, all columns stored by a test will be printed. Default is FALSE and only essential columns related to power will be printed.

Value

A data frame with the number of rows equal to the number of tests. Note that some tests, such as the test by test_parameters(), conduct one test for each parameters. Each such test is counted as one test. The data frame has these columns:

test: The name of the test.

label: The label for each test, or "Test" if a test only does on test (e.g., test_indirect_effect()).

pvalid: The proportion of valid tests across all replications.

reject: The rejection rate for each test. If the null hypothesis is false, then this is the power.

Details

It loops over the tests stored in a power4test object and retrieve the rejection rate of each test.

See also

Examples

model_simple_med <-
"
m ~ x
y ~ m + x
"

model_simple_med_es <- c("y ~ m" = "l",
                         "m ~ x" = "m",
                         "y ~ x" = "n")

sim_only <- power4test(nrep = 4,
                       model = model_simple_med,
                       pop_es = model_simple_med_es,
                       n = 100,
                       R = 50,
                       ci_type = "boot",
                       fit_model_args = list(fit_function = "lm"),
                       do_the_test = FALSE,
                       iseed = 1234)
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Simulate the data:
#> Fit the model(s):
#> Generate bootstrap estimates:

test_out <- power4test(object = sim_only,
                       test_fun = test_indirect_effect,
                       test_args = list(x = "x",
                                        m = "m",
                                        y = "y",
                                        boot_ci = TRUE,
                                        mc_ci = FALSE))
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Do the test: test_indirect: x->m->y 
test_out <- power4test(object = test_out,
                       test_fun = test_parameters)
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Do the test: test_parameters: CIs  
get_rejection_rates(test_out)
#>                     test test_label pvalid reject
#> 1 test_indirect: x->m->y       Test      1   0.75
#> 2  test_parameters: CIs         m~x      1   1.00
#> 3  test_parameters: CIs         y~m      1   1.00
#> 4  test_parameters: CIs         y~x      1   0.00