power4mome: Power Analysis for Moderation and Mediation

Power analysis and sample size determination for moderation, mediation, and moderated mediation in models fitted by structural equation modelling using the 'lavaan' package by Rosseel (2012) doi:10.18637/jss.v048.i02 or by multiple regression. The package 'manymome' by Cheung and Cheung (2024) doi:10.3758/s13428-023-02224-z is used to specify the indirect paths or conditional indirect paths to be tested.

See also

Author

Maintainer: Shu Fai Cheung shufai.cheung@gmail.com (ORCID)

Authors:

Sing-Hang Cheung (ORCID)

Wendie Yang (ORCID)