Estimate the Power of a Test

An all-in-one function that receives a model specification, generates datasets, fits a model, does the target test, and returns the test results.

Usage

power4test(
  object = NULL,
  nrep = NULL,
  ptable = NULL,
  model = NULL,
  pop_es = NULL,
  standardized = TRUE,
  n = NULL,
  number_of_indicators = NULL,
  reliability = NULL,
  x_fun = list(),
  fit_model_args = list(),
  R = NULL,
  ci_type = "mc",
  gen_mc_args = list(),
  gen_boot_args = list(),
  test_fun = NULL,
  test_args = list(),
  map_names = c(fit = "fit"),
  results_fun = NULL,
  results_args = list(),
  test_name = NULL,
  test_note = NULL,
  do_the_test = TRUE,
  sim_all = NULL,
  iseed = NULL,
  parallel = FALSE,
  progress = TRUE,
  ncores = max(1, parallel::detectCores(logical = FALSE) - 1),
  es1 = c(n = 0, nil = 0, s = 0.1, m = 0.3, l = 0.5),
  es2 = c(n = 0, nil = 0, s = 0.05, m = 0.1, l = 0.15),
  n_std = 1090000,
  std_force_monte_carlo = FALSE
)

# S3 method for class 'power4test'
print(
  x,
  what = c("data", "test"),
  digits = 3,
  digits_descriptive = 2,
  data_long = FALSE,
  test_long = FALSE,
  fit_to_all_args = list(),
  ...
)

Arguments

object

Optional. If set to a power4test object, it will be updated using the value(s) in n, pop_es, and/or nrep. Default is NULL.

nrep

The number of replications to generate the simulated datasets. Default is NULL. Must be set when called to create a power4test object.

ptable

The output of ptable_pop(), which is a ptable_pop object, representing the population model. If NULL, the default, ptable_pop() will be called to generate the ptable_pop object using model and pop_es.

model

The lavaan model syntax of the population model. Required. Ignored if ptable is specified. See 'Details' of ptable_pop() on how to use it for models with latent factors and indicators. Ignored if ptable is specified.

pop_es

The character vector to specify population effect sizes. See 'Details' of ptable_pop() on how to set the effect sizes for this argument. Ignored if ptable is specified.

standardized

Logical. If TRUE, the default, variances and error variances are scaled to ensure the population variances of the endogenous variables are close to one, and hence the effect sizes are standardized effect sizes if the variances of the continuos exogenous variables are also equal to one.

n

The sample size for each dataset. Default is 100.

number_of_indicators

A named vector to specify the number of indicators for each factors. See 'Details' of sim_data() on how to set this argument. Default is NULL and all variables in the model syntax are observed variables.

reliability

A named vector to set the reliability coefficient of each set of indicators. Default is NULL. See 'Details' of sim_data() on how to set this argument.

x_fun

The function(s) used to generate the exogenous variables. If not supplied, or set to list(), the default, the variables are generated from a multivariate normal distribution. See 'Details' of sim_data() on how to use this argument.

fit_model_args

A list of the arguments to be passed to fit_model() when fitting the model. Should be a named argument with names being the names of the arguments.

R

The number of replications to generate the Monte Carlo or bootstrapping estimates for each fit output. No Monte Carlo nor bootstrapping estimates will be generated if R is set to NULL.

ci_type

The type of simulation-based confidence intervals to use. Can be either "mc" for Monte Carlo method (the default) or "boot" for nonparametric bootstrapping method. See sim_data() on the details.

gen_mc_args

A list of arguments to be passed to manymome::do_mc() when generating the Monte Carlo estimates. Should be a named argument with names being the names of the arguments. Used only if ci_type is "mc".

gen_boot_args

A list of arguments to be passed to manymome::do_boot() when generating the bootstrap estimates. Should be a named argument with names being the names of the arguments. Used only if ci_type is `"boot".

test_fun

A function to do the test. See the help page of do_test() for the requirements of this function.

test_args

A list of arguments to be passed to the test_fun function. Default is list().

map_names

A named character vector specifying how the content of the element extra in each replication of sim_all map to the argument of test_fun. Default is c(fit = "fit"), indicating that the element fit in the element extra is set to the argument fit of test_fun. That is, for the first replication, fit = sim_out[[1]]$extra$fit when calling test_fun.

results_fun

The function to be used to extract the test results. See Details of do_test() for the requirements of this function. Default is NULL, assuming that the output of test_fun can be used directly.

results_args

A list of arguments to be passed to the results_fun function. Default is list().

test_name

String. The name of the test. Default is NULL, and the name will be created from test_fun. Note that if sim_out is a power4test object and already has a test of this name stored, it will be replaced by the new results.

test_note

String. An optional note for the test, stored in the attribute test_note of the output of do_test(). Default is NULL.

do_the_test

If TRUE, do_test() will be called to do the specified test in the fit output of each dataset.

sim_all

If set to either a sim_out object (the output of sim_out() or a power4test object (the output of power4test()), the stored datasets and fit outputs will be used for doing the test.

iseed

The seed for the random number generator. Default is NULL and the seed is not changed. This seed will be set only once, when calling sim_data().

parallel

If TRUE, parallel processing will be used when calling other functions, if appropriate. Default is FALSE.

progress

If TRUE, the progress of each step will be displayed. Default is `FALSE.

ncores

The number of CPU cores to use if parallel processing is used.

es1

A named vector to set the values for each label of the effect size of correlations and regression paths. Default is c("n" = .00, "nil" = .00, "s" = .10, "m" = .30, "l" = .50). Used only if pop_es is a named vector. See ptable_pop() for further information.

es2

A named vector to set the values for each label of the effect size of product term. Default is c("n" = .00, "nil" = .00, "s" = .05, "m" = .10, "l" = .15). Used only if pop_es is a named vector. See ptable_pop() for further information.

n_std

The sample size used to determine the error variances by simulation when std_force_monte_carlo is TRUE. Default is 100000.

std_force_monte_carlo

Logical. If FALSE, the default, standardization is done analytically if the model has no product terms, and by simulation if the model has product terms. If TRUE, simulation will be used whether the model has product terms or not. Always fall back to standardization if analytical standardization failed.

x

The object to be printed.

what

A string vector of what to print, "data" for simulated data and "test" for stored test(s). Default is c("data", "test").

digits

The numbers of digits displayed after the decimal.

digits_descriptive

The number of digits displayed after the decimal for the descriptive statistics table.

data_long

If TRUE, detailed results will be printed when printing the simulated data.

test_long

If TRUE, detailed results will be printed when printing test(s).

fit_to_all_args

A named list of arguments to be passed to lavaan::sem() when the model is fitted to a sample combined from

...

Optional arguments to be passed to other print methods.

Value

An object of the class power4test, which is a list of with two elements:

sim_all: The output of sim_out().

test_all: A named list of the output of do_test(). The names are the values of test_name. This list can have more than one test because a call to power4test() can add new tests to a power4test object.

The print method of power4test returns x invisibly. Called for its side effect.

Details

It is an all-in-one function for estimating the power of a test for a model, given the sample size and effect sizes.

This is the workflow:

If an object with model and data already generated is supplied through sim_all, such as the output of sim_out() or of power4test() with do_the_test set to FALSE, the following steps will be skipped and go directly to doing the test.

Call sim_data() to determine the population model and generate the datasets.

Call fit_model() to fit the model to each of the datasets.

If R is not NULL and ci_type = "mc", call gen_mc() to generate Monte Carlo estimates using manymome::do_mc().

If R is not NULL and ci_type = "boot", call gen_boot() to generate bootstrap estimates using manymome::do_boot().

Merge the results into a sim_out object by calling sim_out().

If do_the_test is FALSE, skip the remaining steps and return a power4test object, which contains only the data generated and optionally the Monte Carlo or bootstrap estimates.

If do_the_test is TRUE, do the test.

do_test() will be called to do the test in the fit output of each datasets.

Return a power4test object which include the output of sim_out and, if do_the_test is TRUE, the output of do_test().

This function is to be used when users are interested only in the power of one or several tests on a particular aspect of the model, such as a parameter, given a specific effect sizes and sample sizes.

Updating a Condition

This function can also be used to update a condition when only some selected aspects changed. For example, without calling this function with all the arguments set just to change the sample size, it can be called by supplying an existing power4test object and set only n to a new sample size. The data and the tests will be updated automatically.

Multiple Models

More than one model can be fitted to each replication. This is done by setting fit_model_args to a named list. The names are the names used to identify the models, and each element is a list of named list of arguments for a model.

For example:

fit_model_args = list(fit = list(),
                      fit2 = list(model = mod2),
                      fit3 = list(model = mod3))

Three models will be fitted. The first model is the model ued to generate the data, named "fit". The second and first models are named "fit2" and "fit3", respectively, with "fit2" fitted with model = mod2 and "fit3" fitted with model = mod3.

If Monte Carlos or bootstrap estimates are to be generated, they will be generated for each model, using the values for their arguments.

Examples

model_simple_med <-
"
m ~ a * x
y ~ b * m + x
ab := a * b
"

model_simple_med_es <- c("y ~ m" = "l",
                         "m ~ x" = "m",
                         "y ~ x" = "n")

out <- power4test(nrep = 50,
                  model = model_simple_med,
                  pop_es = model_simple_med_es,
                  n = 100,
                  test_fun = test_parameters,
                  test_args = list(pars = "a"),
                  iseed = 1234,
                  parallel = FALSE,
                  progress = TRUE)
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Simulate the data:
#> Fit the model(s):
#> Do the test: test_parameters: CIs (pars: a) 

print(out,
      test_long = TRUE)
#> 
#> ====================== Model Information ======================
#> 
#> == Model on Factors/Variables ==
#> 
#> m ~ a * x
#> y ~ b * m + x
#> ab := a * b
#> 
#> == Model on Variables/Indicators ==
#> 
#> m ~ a * x
#> y ~ b * m + x
#> ab := a * b
#> 
#> ====== Population Values ======
#> 
#> Regressions:
#>                    Population
#>   m ~                        
#>     x          (a)    0.300  
#>   y ~                        
#>     m          (b)    0.500  
#>     x                 0.000  
#> 
#> Variances:
#>                    Population
#>    .m                 0.910  
#>    .y                 0.750  
#>     x                 1.000  
#> 
#> Defined Parameters:
#>                    Population
#>     ab                0.150  
#> 
#> ======================= Data Information =======================
#> 
#> Number of Replications:  50 
#> Sample Sizes:  100 
#> 
#> Call print with 'data_long = TRUE' for further information.
#> 
#> ==================== Extra Element(s) Found ====================
#> 
#> - fit
#> 
#> === Element(s) of the First Dataset ===
#> 
#> ============ <fit> ============
#> 
#> lavaan 0.6-19 ended normally after 1 iteration
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                         5
#> 
#>   Number of observations                           100
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> =========== <test_parameters: CIs (pars: a)> ===========
#> 
#> Mean(s) across replication:
#>   test_label lhs op rhs label   est    se     z pvalue  cilo  cihi   sig
#> 1          a   m  ~   x     a 0.307 0.095 3.247  0.040 0.121 0.494 0.860
#> 
#> - The column 'sig' shows the rejection rates.
#> - If the null hypothesis is false, the rate is the power.
#> - Number of valid replications for rejection rate(s): 50 
#> - Proportion of valid replications for rejection rate(s): 1.000 

# Add one more test

out <- power4test(out,
                  test_fun = test_parameters,
                  test_args = list(op = ":="))
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Do the test: test_parameters: CIs (op: :=) 
print(out,
      test_long = TRUE)
#> 
#> ====================== Model Information ======================
#> 
#> == Model on Factors/Variables ==
#> 
#> m ~ a * x
#> y ~ b * m + x
#> ab := a * b
#> 
#> == Model on Variables/Indicators ==
#> 
#> m ~ a * x
#> y ~ b * m + x
#> ab := a * b
#> 
#> ====== Population Values ======
#> 
#> Regressions:
#>                    Population
#>   m ~                        
#>     x          (a)    0.300  
#>   y ~                        
#>     m          (b)    0.500  
#>     x                 0.000  
#> 
#> Variances:
#>                    Population
#>    .m                 0.910  
#>    .y                 0.750  
#>     x                 1.000  
#> 
#> Defined Parameters:
#>                    Population
#>     ab                0.150  
#> 
#> ======================= Data Information =======================
#> 
#> Number of Replications:  50 
#> Sample Sizes:  100 
#> 
#> Call print with 'data_long = TRUE' for further information.
#> 
#> ==================== Extra Element(s) Found ====================
#> 
#> - fit
#> 
#> === Element(s) of the First Dataset ===
#> 
#> ============ <fit> ============
#> 
#> lavaan 0.6-19 ended normally after 1 iteration
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                         5
#> 
#>   Number of observations                           100
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> =========== <test_parameters: CIs (pars: a)> ===========
#> 
#> Mean(s) across replication:
#>   test_label lhs op rhs label   est    se     z pvalue  cilo  cihi   sig
#> 1          a   m  ~   x     a 0.307 0.095 3.247  0.040 0.121 0.494 0.860
#> 
#> - The column 'sig' shows the rejection rates.
#> - If the null hypothesis is false, the rate is the power.
#> - Number of valid replications for rejection rate(s): 50 
#> - Proportion of valid replications for rejection rate(s): 1.000 
#> 
#> =========== <test_parameters: CIs (op: :=)> ===========
#> 
#> Mean(s) across replication:
#>   test_label lhs op rhs label   est    se     z pvalue  cilo  cihi   sig
#> 1         ab  ab := a*b    ab 0.147 0.054 2.635  0.046 0.040 0.254 0.820
#> 
#> - The column 'sig' shows the rejection rates.
#> - If the null hypothesis is false, the rate is the power.
#> - Number of valid replications for rejection rate(s): 50 
#> - Proportion of valid replications for rejection rate(s): 1.000