Power By Sample Sizes

Estimate power for a range of sample sizes.

The method c.power4test_by_n() is used to combine tests from different runs of power4test_by_n().

The function get_rejection_rates_by_n() is used to extract the rejection rates form a get_rejection_rates_by_n object, with sample sizes added to the output.

Usage

power4test_by_n(object, n = NULL, progress = TRUE, ..., by_seed = NULL)

# S3 method for class 'power4test_by_n'
c(..., sort = TRUE)

get_rejection_rates_by_n(object_by_n, all_columns = FALSE)

Arguments

object

A power4test object.

n

A numeric vector of sample sizes for which the simulation will be conducted.

progress

Logical. Whether progress of the simulation will be displayed.

...

For power4test_by_n(), they are arguments to be passed to power4test(). For c.power4test_by_n(), they are power4test_by_n() outputs to be combined together.

by_seed

If set to a number, it will be used to generate the seeds for each call to power4test(). If NULL, the default, then seeds will still be randomly generated but the results cannot be easily reproduced.

sort

WHen combining power4test_by_n objects, whether they will be sorted by sample sizes. Default is TRUE.

object_by_n

A power4test_by_n object, which is an output of power4test_by_n().

all_columns

If TRUE, all columns stored by a test will be printed. Default is FALSE and only essential columns related to power will be printed.

Value

The function power4test_by_n() returns a power4test_by_n object, which is a list of power4test objects, one for each sample size.

The method c.power4test_by_n() returns a power4test_by_n object with all the elements (tests for different sample sizes) combined.

The function get_rejection_rates_by_n() returns a data frame which is similar to the output of get_rejection_rates(), with a column n added for the sample size for each test.

Details

This function regenerates datasets for a set of sample sizes and does the stored tests in each of them.

Optionally, it can also be run on a object with no stored tests. In this case, additional arguments must be set to instruct power4test() on the tests to be conducted.

It is usually used to examine the power over a range of sample sizes.

See also

Examples

model_simple_med <-
"
m ~ x
y ~ m + x
"

model_simple_med_es <- c("y ~ m" = "l",
                         "m ~ x" = "m",
                         "y ~ x" = "n")

sim_only <- power4test(nrep = 2,
                       model = model_simple_med,
                       pop_es = model_simple_med_es,
                       n = 100,
                       R = 40,
                       ci_type = "boot",
                       fit_model_args = list(fit_function = "lm"),
                       do_the_test = FALSE,
                       iseed = 1234)
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Simulate the data:
#> Fit the model(s):
#> Generate bootstrap estimates:

test_out <- power4test(object = sim_only,
                       test_fun = test_indirect_effect,
                       test_args = list(x = "x",
                                        m = "m",
                                        y = "y",
                                        boot_ci = TRUE,
                                        mc_ci = FALSE))
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Do the test: test_indirect: x->m->y 

out <- power4test_by_n(test_out,
                       n = c(100, 110, 120))
#> Updating the simulation for sample size: 100 
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Re-simulate the data:
#> Fit the model(s):
#> Generate bootstrap estimates:
#> Update the test(s):
#> Update test_indirect: x->m->y :
#> Updating the simulation for sample size: 110 
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Re-simulate the data:
#> Fit the model(s):
#> Generate bootstrap estimates:
#> Update the test(s):
#> Update test_indirect: x->m->y :
#> Updating the simulation for sample size: 120 
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Re-simulate the data:
#> Fit the model(s):
#> Generate bootstrap estimates:
#> Update the test(s):
#> Update test_indirect: x->m->y :
out_reject <- get_rejection_rates_by_n(out)
out_reject
#>     n                   test test_label pvalid reject
#> 1 100 test_indirect: x->m->y       Test      1    0.5
#> 2 110 test_indirect: x->m->y       Test      1    1.0
#> 3 120 test_indirect: x->m->y       Test      1    1.0