Generate the Population Model

Generate the complete population model using the model syntax and user-specified effect sizes.

Usage

ptable_pop(
  model = NULL,
  pop_es = NULL,
  es1 = c(n = 0, nil = 0, s = 0.1, m = 0.3, l = 0.5),
  es2 = c(n = 0, nil = 0, s = 0.05, m = 0.1, l = 0.15),
  standardized = TRUE,
  n_std = 1e+05,
  std_force_monte_carlo = FALSE
)

model_matrices_pop(x, ..., drop_list_single_group = TRUE)

Arguments

model

String. The model defined by lavaan model syntax.

pop_es

It can a data frame with these columns: lhs, op, rhs, and pop. The first three columns correspond to those in a lavaan parameter table. The column pop stores the population values. The column es stores the original labels, for reference. It can also be A named character vector. See 'Details' on how to specify this vector.

es1

A named vector to set the values for each label of the effect size of correlations and regression paths. Default is c("n" = .00, "nil" = .00, "s" = .10, "m" = .30, "l" = .50). Used only if pop_es is a named vector.

es2

A named vector to set the values for each label of the effect size of product term. Default is c("n" = .00, "nil" = .00, "s" = .05, "m" = .10, "l" = .15). Used only if pop_es is a named vector.

standardized

Logical. If TRUE, the default, variances and error variances are scaled to ensure the population variances of the endogenous variables are close to one, and hence the effect sizes are standardized effect sizes if the variances of the continuos exogenous variables are also equal to one.

n_std

The sample size used to determine the error variances by simulation when std_force_monte_carlo is TRUE. Default is 100000.

std_force_monte_carlo

Logical. If FALSE, the default, standardization is done analytically if the model has no product terms, and by simulation if the model has product terms. If TRUE, simulation will be used whether the model has product terms or not. Always fall back to standardization if analytical standardization failed.

x

It can be a 'lavaan' model syntax, to be passed to ptable_pop(), or a parameter table with the column start set to the population values, such as the output of ptable_pop().

...

If x is a model syntax, these are arguments to be passed to ptable_pop().

drop_list_single_group

If TRUE and the number of groups is equal to one, the output will be a list of matrices of one group only. Default if TRUE.

Value

The function ptable_pop() returns a lavaan parameter table of the model, with the column start set to the population values.

The function model_matrices_pop() returns a lavaan LISREL-style model matrices (like the output of lavaan::lavInspect() with what set to "free"), with matrix elements set to the population values. If x is the model syntax, it will be stored in the attributes model. If the model is a multigroup model with k groups (k greater than 1), then it returns a list of k lists of lavaan LISREL-style model matrices unless drop_list_single_group is TRUE.

Details

The function ptable_pop() generates a lavaan parameter table that can be used to generate data based on the population values of model parameters.

This function is used by the all-in-one function power4test(). Users usually do not call this function directly.

Setting pop_es to a named vector

If pop_es is specified by a named vector, it must follow the convention below.

The names of the vectors are lavaan names for the selected parameters. For example, m ~ x denotes the path from x to m.

Alternatively, the names can be either ".beta." or ".cov.". Use ".beta." to set the default values for all regression coefficients. Use ".cov." to set the default values for all correlations of exogenous variables (e.g., predictors).

If using lavaan names, can specify more than one parameter using +. For example, y ~ m + x denotes the two paths from m and x to y.

The value of each element can be the label for the effect size: n for nil, s for small, m for medium, and l for large. The value for each label is determined by es1 and es2.

The value of pop_es can also be set to a value, but it must be quoted as a string, such as "y ~ x" = ".31".

The vector es1 is for correlations and regression coefficients, and the vector es2 is for for standardized moderation effect, the coefficients of a product term.

This is an example:

c(".beta." = "s",
  "m1 ~ x" = "-m",
  "m2 ~ m1" = "l",
  "y ~ x:w" = "s")

In this example,

All regression coefficients are set to small (s) by default, unless specified otherwise.

The path from x to m1 is set to medium and negative (-m).

The path from m1 to m2 is set to large (l).

The coefficient of the product term x:w when predicting y is set to small (s).

Using One Single String To Set pop_es

An alternative way to set pop_es is using one single string, like a lavaan model syntax.

This is an example, equivalent to the example above:

"
.beta.: s
m1 ~ x: -m
m2 ~ m1: l
y ~ x:w: s
"

See the help page of pop_es_yaml() on how to use this approach.

Multigroup Models

The function also supports multigroup models.

Because the model is the population model, equality constraints are irrelevant and the model syntax specifies only the form of the model. Therefore, model is specified as in the case of single group models.

For pop_es, instead of using a named vectors, use as named list.

The names are the parameters, or keywords such as .beta. and .cov., like specifying the population values for a single group model.

The elements are character vectors. If it has only one element (e.g., a single string), then it is the the population value for all groups. If it has more than one element (e.g., a vector of three strings), then they are the population values of the groups. For a model of k groups, each vector must have either k elements or one element.

This is an example:

list("m ~ x" = "m",
     "y ~ m" = c("s", "m", "l"))

In this model, the population value of the path m ~ x is medium for all groups, while the population values for the path y ~ m are small, medium, and large, respectively.

Using One Single String To Set pop_es

The population values for a multigroup model can also be set using one single string.

This is an example, equivalent to the example above:

"
m ~ x: m
y ~ m: [s, m, l]
"

This is another equivalent form:

"
m ~ x: m
y ~ m:
 - s
 - m
 - l
"

See the help page of pop_es_yaml() on how to use this approach.

The function model_matrices_pop() generates models matrices with population values. For advanced users.

Functions

model_matrices_pop():

See also

Examples

model1 <-
"
m1 ~ x + c1
m2 ~ m1 + x2 + c1
y ~  m2 + m1 + x + w + x:w + c1
"

model1_es <- c("m1 ~ x" = "-m",
               "m2 ~ m1" = "s",
               "y ~ m2" = "l",
               "y ~ x" = "m",
               "y ~ w" = "s",
               "y ~ x:w" = "s",
               "x ~~ w" = "s")

ptable_final1 <- ptable_pop(model1,
                            pop_es = model1_es)
ptable_final1
#>    id lhs op rhs group user block free ustart exo label plabel      start est
#> 1   1  m1  ~   x     1    1     1    1     NA   0         .p1. -0.3000000   0
#> 2   2  m1  ~  c1     1    1     1    2     NA   0         .p2.  0.0000000   0
#> 3   3  m2  ~  m1     1    1     1    3     NA   0         .p3.  0.1000000   0
#> 4   4  m2  ~  x2     1    1     1    4     NA   0         .p4.  0.0000000   0
#> 5   5  m2  ~  c1     1    1     1    5     NA   0         .p5.  0.0000000   0
#> 6   6   y  ~  m2     1    1     1    6     NA   0         .p6.  0.5000000   0
#> 7   7   y  ~  m1     1    1     1    7     NA   0         .p7.  0.0000000   0
#> 8   8   y  ~   x     1    1     1    8     NA   0         .p8.  0.3000000   0
#> 9   9   y  ~   w     1    1     1    9     NA   0         .p9.  0.1000000   0
#> 10 10   y  ~ x:w     1    1     1   10     NA   0        .p10.  0.0500000   0
#> 11 11   y  ~  c1     1    1     1   11     NA   0        .p11.  0.0000000   0
#> 12 12  m1 ~~  m1     1    0     1   12     NA   0        .p12.  0.9095567   1
#> 13 13  m2 ~~  m2     1    0     1   13     NA   0        .p13.  0.9899971   1
#> 14 14   y ~~   y     1    0     1   14     NA   0        .p14.  0.6497474   1
#> 15 15   x ~~   x     1    0     1   15     NA   0        .p15.  1.0000000   1
#> 16 16   x ~~  c1     1    0     1   16     NA   0        .p16.  0.0000000   0
#> 17 17   x ~~  x2     1    0     1   17     NA   0        .p17.  0.0000000   0
#> 18 18   x ~~   w     1    0     1   18     NA   0        .p18.  0.1000000   0
#> 19 19   x ~~ x:w     1    0     1   19     NA   0        .p19.  0.0000000   0
#> 20 20  c1 ~~  c1     1    0     1   20     NA   0        .p20.  1.0000000   1
#> 21 21  c1 ~~  x2     1    0     1   21     NA   0        .p21.  0.0000000   0
#> 22 22  c1 ~~   w     1    0     1   22     NA   0        .p22.  0.0000000   0
#> 23 23  c1 ~~ x:w     1    0     1   23     NA   0        .p23.  0.0000000   0
#> 24 24  x2 ~~  x2     1    0     1   24     NA   0        .p24.  1.0000000   1
#> 25 25  x2 ~~   w     1    0     1   25     NA   0        .p25.  0.0000000   0
#> 26 26  x2 ~~ x:w     1    0     1   26     NA   0        .p26.  0.0000000   0
#> 27 27   w ~~   w     1    0     1   27     NA   0        .p27.  1.0000000   1
#> 28 28   w ~~ x:w     1    0     1   28     NA   0        .p28.  0.0000000   0
#> 29 29 x:w ~~ x:w     1    0     1   29     NA   0        .p29.  1.0000000   1

model_matrices_pop(ptable_final1)
#> $lambda
#>     m1 m2 y x c1 x2 w x:w
#> m1   0  0 0 0  0  0 0   0
#> m2   0  0 0 0  0  0 0   0
#> y    0  0 0 0  0  0 0   0
#> x    0  0 0 0  0  0 0   0
#> c1   0  0 0 0  0  0 0   0
#> x2   0  0 0 0  0  0 0   0
#> w    0  0 0 0  0  0 0   0
#> x:w  0  0 0 0  0  0 0   0
#> 
#> $theta
#>     m1 m2 y x c1 x2 w x:w
#> m1   0                   
#> m2   0  0                
#> y    0  0 0              
#> x    0  0 0 0            
#> c1   0  0 0 0  0         
#> x2   0  0 0 0  0  0      
#> w    0  0 0 0  0  0 0    
#> x:w  0  0 0 0  0  0 0   0
#> 
#> $psi
#>       m1   m2    y    x   c1   x2    w  x:w
#> m1  0.91                                   
#> m2  0.00 0.99                              
#> y   0.00 0.00 0.65                         
#> x   0.00 0.00 0.00 1.00                    
#> c1  0.00 0.00 0.00 0.00 1.00               
#> x2  0.00 0.00 0.00 0.00 0.00 1.00          
#> w   0.00 0.00 0.00 0.10 0.00 0.00 1.00     
#> x:w 0.00 0.00 0.00 0.00 0.00 0.00 0.00 1.00
#> 
#> $beta
#>      m1  m2 y    x c1 x2   w  x:w
#> m1  0.0 0.0 0 -0.3  0  0 0.0 0.00
#> m2  0.1 0.0 0  0.0  0  0 0.0 0.00
#> y   0.0 0.5 0  0.3  0  0 0.1 0.05
#> x   0.0 0.0 0  0.0  0  0 0.0 0.00
#> c1  0.0 0.0 0  0.0  0  0 0.0 0.00
#> x2  0.0 0.0 0  0.0  0  0 0.0 0.00
#> w   0.0 0.0 0  0.0  0  0 0.0 0.00
#> x:w 0.0 0.0 0  0.0  0  0 0.0 0.00
#> 
#> attr(,"model")
#> [1] "\nm1 ~ x + c1\nm2 ~ m1 + x2 + c1\ny ~  m2 + m1 + x + w + x:w + c1\n"

model_matrices_pop(model1,
                   pop_es = model1_es)
#> $lambda
#>     m1 m2 y x c1 x2 w x:w
#> m1   0  0 0 0  0  0 0   0
#> m2   0  0 0 0  0  0 0   0
#> y    0  0 0 0  0  0 0   0
#> x    0  0 0 0  0  0 0   0
#> c1   0  0 0 0  0  0 0   0
#> x2   0  0 0 0  0  0 0   0
#> w    0  0 0 0  0  0 0   0
#> x:w  0  0 0 0  0  0 0   0
#> 
#> $theta
#>     m1 m2 y x c1 x2 w x:w
#> m1   0                   
#> m2   0  0                
#> y    0  0 0              
#> x    0  0 0 0            
#> c1   0  0 0 0  0         
#> x2   0  0 0 0  0  0      
#> w    0  0 0 0  0  0 0    
#> x:w  0  0 0 0  0  0 0   0
#> 
#> $psi
#>        m1    m2     y     x    c1    x2     w   x:w
#> m1  0.910                                          
#> m2  0.000 0.990                                    
#> y   0.000 0.000 0.652                              
#> x   0.000 0.000 0.000 1.000                        
#> c1  0.000 0.000 0.000 0.000 1.000                  
#> x2  0.000 0.000 0.000 0.000 0.000 1.000            
#> w   0.000 0.000 0.000 0.100 0.000 0.000 1.000      
#> x:w 0.000 0.000 0.000 0.000 0.000 0.000 0.000 1.000
#> 
#> $beta
#>      m1  m2 y    x c1 x2   w  x:w
#> m1  0.0 0.0 0 -0.3  0  0 0.0 0.00
#> m2  0.1 0.0 0  0.0  0  0 0.0 0.00
#> y   0.0 0.5 0  0.3  0  0 0.1 0.05
#> x   0.0 0.0 0  0.0  0  0 0.0 0.00
#> c1  0.0 0.0 0  0.0  0  0 0.0 0.00
#> x2  0.0 0.0 0  0.0  0  0 0.0 0.00
#> w   0.0 0.0 0  0.0  0  0 0.0 0.00
#> x:w 0.0 0.0 0  0.0  0  0 0.0 0.00
#> 
#> attr(,"model")
#> [1] "\nm1 ~ x + c1\nm2 ~ m1 + x2 + c1\ny ~  m2 + m1 + x + w + x:w + c1\n"