Summarize 'n_from_power' Results

The summary method of the output of n_from_power().

Usage

# S3 method for class 'n_from_power'
summary(object, ...)

# S3 method for class 'summary.n_from_power'
print(x, digits = 3, ...)

Arguments

object

An n_from_power-class object, such as the output of n_from_power().

...

Additional arguments. Not used for now.

x

The output of summary.n_from_power(), the summary method of an n_from_power object, which is the output of n_from_power().

digits

The number of digits after the decimal when printing the results.

Value

Return an object of the class summary.n_from_power, which is simply the output of n_from_power(), with a print method dedicated for detailed summary. Please refer to n_from_power() for the contents.

The print-method of summary.n_from_power objects returns the object x invisibly. It is called for its side effect.

Details

It simply prepares the results of n_from_power() to be printed in details.

See also

Examples

mod <-
"
m ~ a*x
y ~ b*m + x
ab := a * b
"

mod_es <- c("y ~ m" = "l",
            "m ~ x" = "m",
            "y ~ x" = "n")

sim_only <- power4test(nrep = 10,
                       model = mod,
                       pop_es = mod_es,
                       n = 100,
                       do_the_test = FALSE,
                       iseed = 1234)
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Simulate the data:
#> Fit the model(s):

test_out <- power4test(object = sim_only,
                       test_fun = test_parameters,
                       test_args = list(pars = "ab"))
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Do the test: test_parameters: CIs (pars: ab) 

# In real analysis, to have more stable results:
# - Use a larger final_nrep (e.g., 500).
# - Use the default ns_per_trial of 3, or just remove it.

# If the default values are OK, this call is sufficient:
# power_vs_n <- n_from_power(test_out,
#                            target_power = .80,
#                            seed = 4567)
power_vs_n <- n_from_power(test_out,
                           progress = TRUE,
                           target_power = .80,
                           final_nrep = 10,
                           ns_per_trial = 1,
                           nrep_steps = 1,
                           max_trials = 1,
                           seed = 4567)
#> 
#> --- Pre-iteration Search ---
#> 
#> - Start at 2025-05-13 12:40:14 
#> - Sample sizes to try:  125 
#> Updating the simulation for sample size: 125 
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Re-simulate the data:
#> Fit the model(s):
#> Update the test(s):
#> Update test_parameters: CIs (pars: ab) :
#> - Rejection Rates:
#>     n                            test test_label pvalid reject
#> 1 100 test_parameters: CIs (pars: ab)         ab      1    0.8
#> 2 125 test_parameters: CIs (pars: ab)         ab      1    1.0
#> 
#> - 'nls()' estimation skipped when less than 4 sample sizes examined.
#> - Power Curve:
#> 
#> Call:  stats::glm(formula = sig ~ n, family = "binomial", data = reject1)
#> 
#> Coefficients:
#> (Intercept)            n  
#>    -75.3328       0.7672  
#> 
#> Degrees of Freedom: 19 Total (i.e. Null);  18 Residual
#> Null Deviance:	    13 
#> Residual Deviance: 10.01 	AIC: 14.01
#> 
#> 
#> 
#> --- Trial 1 ---
#> 
#> - Start at 2025-05-13 12:40:14 
#> - Sample sizes to try: 98, 60 
#> - Numbers of replications: 10, 10 
#> Updating the simulation for sample size: 98 
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Re-simulate the data:
#> Fit the model(s):
#> Update the test(s):
#> Update test_parameters: CIs (pars: ab) :
#> Updating the simulation for sample size: 60 
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Re-simulate the data:
#> Fit the model(s):
#> Update the test(s):
#> Update test_parameters: CIs (pars: ab) :
#> - Rejection Rates:
#>     n                            test test_label pvalid reject
#> 1  60 test_parameters: CIs (pars: ab)         ab      1    0.3
#> 2  98 test_parameters: CIs (pars: ab)         ab      1    0.8
#> 3 100 test_parameters: CIs (pars: ab)         ab      1    0.8
#> 4 125 test_parameters: CIs (pars: ab)         ab      1    1.0
#> 
#> - 'nls()' estimation failed. Switch to logistic regression.
#> - Sample size with closest power: 98 
#> - Estimated power for 98: 0.8000
#> - Power Curve:
#> 
#> Call:  stats::glm(formula = sig ~ n, family = "binomial", data = reject1)
#> 
#> Coefficients:
#> (Intercept)            n  
#>    -4.88278      0.06534  
#> 
#> Degrees of Freedom: 39 Total (i.e. Null);  38 Residual
#> Null Deviance:	    47.05 
#> Residual Deviance: 33.13 	AIC: 37.13
#> 
#> - Estimated power is close enough to target power (0.8000). (CI: [0.5521,1.0479])
#> 
#> 
#> --- Final Stage ---
#> 
#> - Start at 2025-05-13 12:40:16 
#> - Rejection Rates:
#>     n                            test test_label pvalid reject
#> 1  60 test_parameters: CIs (pars: ab)         ab      1    0.3
#> 2  98 test_parameters: CIs (pars: ab)         ab      1    0.8
#> 3 100 test_parameters: CIs (pars: ab)         ab      1    0.8
#> 4 125 test_parameters: CIs (pars: ab)         ab      1    1.0
#> 
#> - Estimated Power Curve:
#> 
#> Call:  stats::glm(formula = sig ~ n, family = "binomial", data = reject1)
#> 
#> Coefficients:
#> (Intercept)            n  
#>    -4.88278      0.06534  
#> 
#> Degrees of Freedom: 39 Total (i.e. Null);  38 Residual
#> Null Deviance:	    47.05 
#> Residual Deviance: 33.13 	AIC: 37.13
#> 
#> 
#> - Final Sample Size: 98 
#> - Final Estimated Power: 0.8000 
#> - Confidence Interval: [0.5521; 1.0479]
#> - CI Level: 95.00%
summary(power_vs_n)
#> 
#> ====== n_from_power Results ======
#> 
#> Call:
#> n_from_power(object = test_out, target_power = 0.8, ns_per_trial = 1, 
#>     progress = TRUE, max_trials = 1, final_nrep = 10, nrep_steps = 1, 
#>     seed = 4567)
#> 
#> - Target Power: 0.800 
#> 
#> === Major Results ===
#> 
#> - Final Sample Size: 98 
#> - Final Estimated Power: 0.800 
#> - Confidence Interval: [0.552; 1.048]
#> - Level of confidence: 95.0%
#> - Based on 10 replications.
#> 
#> === Technical Information ===
#> 
#> - The range of sample sizes explored: 60 to 125 
#> - Time spent in the search: 2.082 secs 
#> - The final crude model for the power-sample-size relation:
#> 
#> Model Type: Logistic Regression 
#> 
#> 
#> Call:  stats::glm(formula = sig ~ n, family = "binomial", data = reject1)
#> 
#> Coefficients:
#> (Intercept)            n  
#>    -4.88278      0.06534  
#> 
#> Degrees of Freedom: 39 Total (i.e. Null);  38 Residual
#> Null Deviance:	    47.05 
#> Residual Deviance: 33.13 	AIC: 37.13
#> 
#> - Detailed Results:
#> 
#>     n                            test test_label pvalid reject
#> 1  60 test_parameters: CIs (pars: ab)         ab      1    0.3
#> 2  98 test_parameters: CIs (pars: ab)         ab      1    0.8
#> 3 100 test_parameters: CIs (pars: ab)         ab      1    0.8
#> 4 125 test_parameters: CIs (pars: ab)         ab      1    1.0
#>