Summarize 'x_from_power' Results

The summary method of the output of x_from_power().

Usage

# S3 method for class 'x_from_power'
summary(object, ...)

# S3 method for class 'summary.x_from_power'
print(x, digits = 3, ...)

Arguments

object

An x_from_power-class object, such as the output of x_from_power().

...

Additional arguments. Not used for now.

x

The output of summary.x_from_power(), the summary method of an x_from_power object, which is the output of x_from_power().

digits

The number of digits after the decimal when printing the results.

Value

Return an object of the class summary.x_from_power, which is simply the output of x_from_power(), with a print method dedicated for detailed summary. Please refer to x_from_power() for the contents.

The print-method of summary.x_from_power objects returns the object x invisibly. It is called for its side effect.

Details

It simply prepares the results of x_from_power() to be printed in details.

See also

Examples

mod <-
"
m ~ a*x
y ~ b*m + x
ab := a * b
"

mod_es <- c("y ~ m" = "l",
            "m ~ x" = "m",
            "y ~ x" = "n")

sim_only <- power4test(nrep = 10,
                       model = mod,
                       pop_es = mod_es,
                       n = 100,
                       do_the_test = FALSE,
                       iseed = 1234)
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Simulate the data:
#> Fit the model(s):

test_out <- power4test(object = sim_only,
                       test_fun = test_parameters,
                       test_args = list(pars = "ab"))
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Do the test: test_parameters: CIs (pars: ab) 

# In real analysis, to have more stable results:
# - Use a larger final_nrep (e.g., 500).
# - Use the default xs_per_trial of 3, or just remove it.

# If the default values are OK, this call is sufficient:
# power_vs_n <- x_from_power(test_out,
#                            x = "n",
#                            target_power = .80,
#                            seed = 4567)
power_vs_n <- x_from_power(test_out,
                           x = "n",
                           progress = TRUE,
                           target_power = .80,
                           final_nrep = 10,
                           xs_per_trial = 1,
                           nrep_steps = 1,
                           max_trials = 1,
                           seed = 4567)
#> 
#> --- Pre-iteration Search ---
#> 
#> - Start at 2025-05-18 12:47:53 
#> - Value(s) to try:  125 
#> Updating the simulation for sample size: 125 
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Re-simulate the data:
#> Fit the model(s):
#> Update the test(s):
#> Update test_parameters: CIs (pars: ab) :
#> - Rejection Rates:
#>     n                            test test_label pvalid reject reject_ci_lo
#> 1 100 test_parameters: CIs (pars: ab)         ab      1    0.8     0.552082
#> 2 125 test_parameters: CIs (pars: ab)         ab      1    1.0     1.000000
#>   reject_ci_hi
#> 1     1.047918
#> 2     1.000000
#> 
#> - 'nls()' estimation skipped when less than 4 values of predictor examined.
#> - Power Curve:
#> Call:
#> power_curve(object = by_x_i, formula = power_model, start = start, 
#>     lower_bound = lower_bound, upper_bound = upper_bound, nls_args = nls_args, 
#>     nls_control = nls_control, verbose = progress)
#> 
#> Predictor: n (Sample Size)
#> 
#> Model:
#> 
#> Call:  stats::glm(formula = reject ~ x, family = "binomial", data = reject1)
#> 
#> Coefficients:
#> (Intercept)            x  
#>    -75.3328       0.7672  
#> 
#> Degrees of Freedom: 19 Total (i.e. Null);  18 Residual
#> Null Deviance:	    13 
#> Residual Deviance: 10.01 	AIC: 14.01
#> 
#> 
#> 
#> --- Trial 1 ---
#> 
#> - Start at 2025-05-18 12:47:54 
#> - Value(s) to try: 98, 101 
#> - Numbers of replications: 10, 7 
#> Updating the simulation for sample size: 98 
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Re-simulate the data:
#> Fit the model(s):
#> Update the test(s):
#> Update test_parameters: CIs (pars: ab) :
#> Updating the simulation for sample size: 101 
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Re-simulate the data:
#> Fit the model(s):
#> Update the test(s):
#> Update test_parameters: CIs (pars: ab) :
#> - Rejection Rates:
#>     n                            test test_label pvalid    reject reject_ci_lo
#> 1  98 test_parameters: CIs (pars: ab)         ab      1 0.8000000     0.552082
#> 2 100 test_parameters: CIs (pars: ab)         ab      1 0.8000000     0.552082
#> 3 101 test_parameters: CIs (pars: ab)         ab      1 0.8571429     0.597918
#> 4 125 test_parameters: CIs (pars: ab)         ab      1 1.0000000     1.000000
#>   reject_ci_hi
#> 1     1.047918
#> 2     1.047918
#> 3     1.116368
#> 4     1.000000
#> 
#> - 'nls()' estimation failed. Switch to logistic regression.
#> Updating the simulation for sample size: 99 
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Re-simulate the data:
#> Fit the model(s):
#> Update the test(s):
#> Update test_parameters: CIs (pars: ab) :
#> - Extrapolated value: 99 
#> - Estimated power for 99: 0.7000
#> - 'nls()' estimation failed. Switch to logistic regression.
#> - Power Curve:
#> Call:
#> power_curve(object = by_x_1, formula = power_model, start = stats::coef(fit_1), 
#>     lower_bound = lower_bound, upper_bound = upper_bound, nls_args = nls_args, 
#>     nls_control = nls_control, verbose = progress)
#> 
#> Predictor: n (Sample Size)
#> 
#> Model:
#> 
#> Call:  stats::glm(formula = reject ~ x, family = "binomial", data = reject1)
#> 
#> Coefficients:
#> (Intercept)            x  
#>    -18.6662       0.2009  
#> 
#> Degrees of Freedom: 46 Total (i.e. Null);  45 Residual
#> Null Deviance:	    42.89 
#> Residual Deviance: 38.54 	AIC: 42.54
#> 
#> - Estimated power is close enough to target power (0.8000). (CI: [0.5521,1.0479])
#> 
#> 
#> --- Final Stage ---
#> 
#> - Start at 2025-05-18 12:47:55 
#> - Rejection Rates:
#>     n                            test test_label pvalid    reject reject_ci_lo
#> 1  98 test_parameters: CIs (pars: ab)         ab      1 0.8000000    0.5520820
#> 2  99 test_parameters: CIs (pars: ab)         ab      1 0.7000000    0.4159742
#> 3 100 test_parameters: CIs (pars: ab)         ab      1 0.8000000    0.5520820
#> 4 101 test_parameters: CIs (pars: ab)         ab      1 0.8571429    0.5979180
#> 5 125 test_parameters: CIs (pars: ab)         ab      1 1.0000000    1.0000000
#>   reject_ci_hi
#> 1    1.0479180
#> 2    0.9840258
#> 3    1.0479180
#> 4    1.1163677
#> 5    1.0000000
#> 
#> - Estimated Power Curve:
#> Call:
#> power_curve(object = by_x_i, formula = power_model, start = start, 
#>     lower_bound = lower_bound, upper_bound = upper_bound, nls_args = nls_args, 
#>     nls_control = nls_control, verbose = progress)
#> 
#> Predictor: n (Sample Size)
#> 
#> Model:
#> 
#> Call:  stats::glm(formula = reject ~ x, family = "binomial", data = reject1)
#> 
#> Coefficients:
#> (Intercept)            x  
#>    -75.3328       0.7672  
#> 
#> Degrees of Freedom: 19 Total (i.e. Null);  18 Residual
#> Null Deviance:	    13 
#> Residual Deviance: 10.01 	AIC: 14.01
#> 
#> 
#> - Final Value: 98 
#> - Final Estimated Power: 0.8000 
#> - Confidence Interval: [0.5521; 1.0479]
#> - CI Level: 95.00%
summary(power_vs_n)
#> 
#> ====== x_from_power Results ======
#> 
#> Call:
#> x_from_power(object = test_out, x = "n", target_power = 0.8, 
#>     xs_per_trial = 1, progress = TRUE, max_trials = 1, final_nrep = 10, 
#>     nrep_steps = 1, seed = 4567)
#> Predictor (x): Sample Size 
#> 
#> - Target Power: 0.800 
#> 
#> === Major Results ===
#> 
#> - Final Value: 98 
#> - Final Estimated Power: 0.800 
#> - Confidence Interval: [0.552; 1.048]
#> - Level of confidence: 95.0%
#> - Based on 10 replications.
#> 
#> === Technical Information ===
#> 
#> - The range of values explored: 98 to 125 
#> - Time spent in the search: 2.27 secs 
#> - The final crude model for the power-predictor relation:
#> 
#> Model Type: Logistic Regression 
#> 
#> Call:
#> power_curve(object = by_x_i, formula = power_model, start = start, 
#>     lower_bound = lower_bound, upper_bound = upper_bound, nls_args = nls_args, 
#>     nls_control = nls_control, verbose = progress)
#> 
#> Predictor: n (Sample Size)
#> 
#> Model:
#> 
#> Call:  stats::glm(formula = reject ~ x, family = "binomial", data = reject1)
#> 
#> Coefficients:
#> (Intercept)            x  
#>    -75.3328       0.7672  
#> 
#> Degrees of Freedom: 19 Total (i.e. Null);  18 Residual
#> Null Deviance:	    13 
#> Residual Deviance: 10.01 	AIC: 14.01
#> 
#> - Detailed Results:
#> 
#>     n                            test test_label pvalid reject reject_ci_lo
#> 1  98 test_parameters: CIs (pars: ab)         ab      1  0.800        0.552
#> 2  99 test_parameters: CIs (pars: ab)         ab      1  0.700        0.416
#> 3 100 test_parameters: CIs (pars: ab)         ab      1  0.800        0.552
#> 4 101 test_parameters: CIs (pars: ab)         ab      1  0.857        0.598
#> 5 125 test_parameters: CIs (pars: ab)         ab      1  1.000        1.000
#>   reject_ci_hi
#> 1        1.048
#> 2        0.984
#> 3        1.048
#> 4        1.116
#> 5        1.000
#>