Test All Moderation Effects

Test all moderation effects by testing all product terms for a power4test object.

Usage

test_moderation(
  fit = fit,
  standardized = FALSE,
  ...,
  fit_name = "fit",
  get_map_names = FALSE,
  get_test_name = FALSE
)

Arguments

fit

The fit object, to be passed to lavaan::parameterEstimates(), lavaan::standardizedSolution(), or lmhelprs::lm_list_to_partable().

standardized

Logical. If TRUE, lavaan::standardizedSolution() will be used. Can be used only with models fitted by lavaan.

...

Additional arguments to be passed to lavaan::parameterEstimates(), lavaan::standardizedSolution(), or lmhelprs::lm_list_to_partable().

fit_name

The name of the model fit object to be extracted. Default is "fit". Used only when more than one model is fitted in each replication. This should be the name of the model on which the test is to be conducted.

get_map_names

Logical. Used by power4test() to determine how to extract stored information and assign them to this function. Users should not use this argument.

get_test_name

Logical. Used by power4test() to get the default name of this test. Users should not use this argument.

Value

In its normal usage, it returns the output returned by lavaan::parameterEstimates() or lmhelprs::lm_list_to_partable(), with the following modifications:

est: The parameter estimates, even if standardized estimates are requested (not est.std).

cilo and cihi: The lower and upper limits of the confidence interval (95% by default), respectively (not ci.lower and ci.upper).

sig: Whether a test by confidence interval is significant (1) or not significant (0).

test_label: A column of labels generated by lavaan::lav_partable_labels(), which are usually the labels used by coef() to label the parameters.

Details

This function is to be used in power4test() for testing all product terms, by setting it to the test_fun argument.

It is just a wrapper to test_parameters(). It will first identifies all product terms (terms with : in the names), and then call test_parameters(), with pars set to select the regression coefficients of these terms.

See also

Examples

mod <-
"
m ~ x + w1 + x:w1
y ~ m + w2 + m:w2 + x
"

mod_es <- c("m ~ x" = "n",
            "y ~ x" = "m",
            "m ~ w1" = "n",
            "m ~ x:w1" = "l",
            "y ~ m:w2" = "-s")

sim_only <- power4test(nrep = 5,
                       model = mod,
                       pop_es = mod_es,
                       n = 100,
                       do_the_test = FALSE,
                       iseed = 1234)
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Simulate the data:
#> Fit the model(s):

test_out <- power4test(object = sim_only,
                       test_fun = test_moderation)
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Do the test: test_moderation: CIs  

print(test_out,
      test_long = TRUE)
#> 
#> ====================== Model Information ======================
#> 
#> == Model on Factors/Variables ==
#> 
#> m ~ x + w1 + x:w1
#> y ~ m + w2 + m:w2 + x
#> 
#> == Model on Variables/Indicators ==
#> 
#> m ~ x + w1 + x:w1
#> y ~ m + w2 + m:w2 + x
#> 
#> ====== Population Values ======
#> 
#> Regressions:
#>                    Population
#>   m ~                        
#>     x                 0.000  
#>     w1                0.000  
#>     x:w1              0.150  
#>   y ~                        
#>     m                 0.000  
#>     w2                0.000  
#>     m:w2             -0.050  
#>     x                 0.300  
#> 
#> Covariances:
#>                    Population
#>   x ~~                       
#>     w1                0.000  
#>     x:w1              0.000  
#>     w2                0.000  
#>   w1 ~~                      
#>     x:w1              0.000  
#>     w2                0.000  
#>   x:w1 ~~                    
#>     w2                0.000  
#> 
#> Variances:
#>                    Population
#>    .m                 0.977  
#>    .y                 0.907  
#>     m:w2              1.000  
#>     x                 1.000  
#>     w1                1.000  
#>     x:w1              1.000  
#>     w2                1.000  
#> 
#> ======================= Data Information =======================
#> 
#> Number of Replications:  5 
#> Sample Sizes:  100 
#> 
#> Call print with 'data_long = TRUE' for further information.
#> 
#> ==================== Extra Element(s) Found ====================
#> 
#> - fit
#> 
#> === Element(s) of the First Dataset ===
#> 
#> ============ <fit> ============
#> 
#> lavaan 0.6-19 ended normally after 9 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        10
#> 
#>   Number of observations                           100
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                                12.032
#>   Degrees of freedom                                 8
#>   P-value (Chi-square)                           0.150
#> 
#> =============== <test_moderation: CIs > ===============
#> 
#> Mean(s) across replication:
#>   test_label lhs op  rhs    est    se      z pvalue   cilo  cihi   sig
#> 1     m~x:w1   m  ~ x:w1  0.178 0.106  1.624  0.204 -0.031 0.386 0.600
#> 2     y~m:w2   y  ~ m:w2 -0.124 0.099 -1.247  0.374 -0.318 0.071 0.200
#> 
#> - The column 'sig' shows the rejection rates.
#> - If the null hypothesis is false, the rate is the power.
#> - Number of valid replications for rejection rate(s): 5 
#> - Proportion of valid replications for rejection rate(s): 1.000