Summarize Test Results

Extract and summarize test results.

Usage

summarize_tests(object)

# S3 method for class 'test_summary_list'
print(x, digits = 3, ...)

# S3 method for class 'test_summary'
print(x, digits = 2, ...)

# S3 method for class 'test_out_list'
print(x, digits = 3, test_long = FALSE, ...)

Arguments

object

A power4test object or the element test_all in a power4test object.

x

The object to be printed.

digits

The numbers of digits after the decimal when print numeric results.

...

Optional arguments. Not used.

test_long

If TRUE, a detailed report will be printed.

Value

The function summarize_tests() returns a list of the class test_summary_list. Each element contains a summary of a test stored. The elements are of the class test_summary, with these elments:

test_attributes: The stored information of a test, for printing.

nrep: The number of datasets (replications).

mean: The means of numeric information. For significance tests, these are the rejection rates.

nvalid: The number of non-NA replications used to compute each mean.

The print methods return x invisibly. They are called for their side effects.

Details

This function is used to extract information from each test stored in a power4test object.

The method print.test_out_list() is used to print the contest of a list of test stored in a power4test object, with the option to print just the names of tests.

See also

Examples

mod <-
"
m ~ a*x
y ~ b*m + x
ab := a * b
"

mod_es <- c("y ~ m" = "l",
            "m ~ x" = "m",
            "y ~ x" = "n")

sim_only <- power4test(nrep = 2,
                       model = mod,
                       pop_es = mod_es,
                       n = 100,
                       do_the_test = FALSE,
                       iseed = 1234)
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Simulate the data:
#> Fit the model(s):

test_out <- power4test(object = sim_only,
                       test_fun = test_parameters)
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Do the test: test_parameters: CIs  

test_out
#> 
#> ====================== Model Information ======================
#> 
#> == Model on Factors/Variables ==
#> 
#> m ~ a*x
#> y ~ b*m + x
#> ab := a * b
#> 
#> == Model on Variables/Indicators ==
#> 
#> m ~ a*x
#> y ~ b*m + x
#> ab := a * b
#> 
#> ====== Population Values ======
#> 
#> Regressions:
#>                    Population
#>   m ~                        
#>     x          (a)    0.300  
#>   y ~                        
#>     m          (b)    0.500  
#>     x                 0.000  
#> 
#> Variances:
#>                    Population
#>    .m                 0.910  
#>    .y                 0.750  
#>     x                 1.000  
#> 
#> Defined Parameters:
#>                    Population
#>     ab                0.150  
#> 
#> ======================= Data Information =======================
#> 
#> Number of Replications:  2 
#> Sample Sizes:  100 
#> 
#> Call print with 'data_long = TRUE' for further information.
#> 
#> ==================== Extra Element(s) Found ====================
#> 
#> - fit
#> 
#> === Element(s) of the First Dataset ===
#> 
#> ============ <fit> ============
#> 
#> lavaan 0.6-19 ended normally after 1 iteration
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                         5
#> 
#>   Number of observations                           100
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> ====================== Test(s) Conducted ======================
#> 
#> - test_parameters: CIs 
#> 
#> Call print() and set 'test_long = TRUE' for a detailed report.