Test All Free Parameters

Test all free parameters, including user-defined parameters, for a power4test object.

Usage

test_parameters(
  fit = fit,
  standardized = FALSE,
  pars = NULL,
  op = NULL,
  remove.nonfree = TRUE,
  ...,
  fit_name = "fit",
  get_map_names = FALSE,
  get_test_name = FALSE
)

Arguments

fit

The fit object, to be passed to lavaan::parameterEstimates(), lavaan::standardizedSolution(), or lmhelprs::lm_list_to_partable().

standardized

Logical. If TRUE, lavaan::standardizedSolution() will be used. Can be used only with models fitted by lavaan.

pars

Optional. If set to a character vector, only parameters with test_label equal to values in pars will be returned.

op

Optional. If set to a character vector, only parameters with operators (e.g., "~", "=~") will be returned. If both pars and op are specified, only parameters meeting both requirements will be returned.

remove.nonfree

Logical. If TRUE, the default, only free parameters will be returned. Ignored if standardized is TRUE or if the model is not fitted by lavaan.

...

Additional arguments to be passed to lavaan::parameterEstimates(), lavaan::standardizedSolution(), or lmhelprs::lm_list_to_partable().

fit_name

The name of the model fit object to be extracted. Default is "fit". Used only when more than one model is fitted in each replication. This should be the name of the model on which the test is to be conducted.

get_map_names

Logical. Used by power4test() to determine how to extract stored information and assign them to this function. Users should not use this argument.

get_test_name

Logical. Used by power4test() to get the default name of this test. Users should not use this argument.

Value

In its normal usage, it returns the output returned by lavaan::parameterEstimates() or lmhelprs::lm_list_to_partable(), with the following modifications:

est: The parameter estimates, even if standardized estimates are requested (not est.std).

cilo and cihi: The lower and upper limits of the confidence interval (95% by default), respectively (not ci.lower and ci.upper).

sig: Whether a test by confidence interval is significant (1) or not significant (0).

test_label: A column of labels generated by lavaan::lav_partable_labels(), which are usually the labels used by coef() to label the parameters.

Details

This function is to be used in power4test() for testing all free and user-defined model parameters, by setting it to the test_fun argument.

For models fitted by lavaan, it uses lavaan::parameterEstimates() to do the test. If bootstrapping was requested (by setting se = "boot"), then it supports bootstrap confidence intervals returned by lavaan::parameterEstimates().

It has preliminary, though limited, supported for models fitted by stats::lm() (through lmhelprs::many_lm()). Tests are conducted by ordinary least squares confidence intervals based on the t statistic, reported by stats::confint() applied to the output of stats::lm().

See also

Examples

mod <-
"
m ~ a*x
y ~ b*m + x
ab := a * b
"

mod_es <- c("y ~ m" = "l",
            "m ~ x" = "m",
            "y ~ x" = "n")

sim_only <- power4test(nrep = 2,
                       model = mod,
                       pop_es = mod_es,
                       n = 100,
                       do_the_test = FALSE,
                       iseed = 1234)
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Simulate the data:
#> Fit the model(s):

test_out <- power4test(object = sim_only,
                       test_fun = test_parameters)
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Do the test: test_parameters: CIs  

print(test_out,
      test_long = TRUE)
#> 
#> ====================== Model Information ======================
#> 
#> == Model on Factors/Variables ==
#> 
#> m ~ a*x
#> y ~ b*m + x
#> ab := a * b
#> 
#> == Model on Variables/Indicators ==
#> 
#> m ~ a*x
#> y ~ b*m + x
#> ab := a * b
#> 
#> ====== Population Values ======
#> 
#> Regressions:
#>                    Population
#>   m ~                        
#>     x          (a)    0.300  
#>   y ~                        
#>     m          (b)    0.500  
#>     x                 0.000  
#> 
#> Variances:
#>                    Population
#>    .m                 0.910  
#>    .y                 0.750  
#>     x                 1.000  
#> 
#> Defined Parameters:
#>                    Population
#>     ab                0.150  
#> 
#> ======================= Data Information =======================
#> 
#> Number of Replications:  2 
#> Sample Sizes:  100 
#> 
#> Call print with 'data_long = TRUE' for further information.
#> 
#> ==================== Extra Element(s) Found ====================
#> 
#> - fit
#> 
#> === Element(s) of the First Dataset ===
#> 
#> ============ <fit> ============
#> 
#> lavaan 0.6-19 ended normally after 1 iteration
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                         5
#> 
#>   Number of observations                           100
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> =============== <test_parameters: CIs > ===============
#> 
#> Mean(s) across replication:
#>   test_label lhs op rhs label    est    se      z pvalue   cilo  cihi   sig
#> 1          a   m  ~   x     a  0.382 0.103  3.698  0.000  0.179 0.585 1.000
#> 2          b   y  ~   m     b  0.443 0.094  4.715  0.000  0.259 0.627 1.000
#> 3        y~x   y  ~   x       -0.029 0.104 -0.226  0.236 -0.232 0.174 0.000
#> 4       m~~m   m ~~   m        0.945 0.134  7.071  0.000  0.683 1.207 1.000
#> 5       y~~y   y ~~   y        0.836 0.118  7.071  0.000  0.604 1.068 1.000
#> 6         ab  ab := a*b    ab  0.169 0.058  2.909  0.004  0.055 0.283 1.000
#> 
#> - The column 'sig' shows the rejection rates.
#> - If the null hypothesis is false, the rate is the power.
#> - Number of valid replications for rejection rate(s): 2 
#> - Proportion of valid replications for rejection rate(s): 1.000 

test_out <- power4test(object = sim_only,
                       test_fun = test_parameters,
                       test_args = list(standardized = TRUE))
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Do the test: test_parameters: CIs (standardized)  

print(test_out,
      test_long = TRUE)
#> 
#> ====================== Model Information ======================
#> 
#> == Model on Factors/Variables ==
#> 
#> m ~ a*x
#> y ~ b*m + x
#> ab := a * b
#> 
#> == Model on Variables/Indicators ==
#> 
#> m ~ a*x
#> y ~ b*m + x
#> ab := a * b
#> 
#> ====== Population Values ======
#> 
#> Regressions:
#>                    Population
#>   m ~                        
#>     x          (a)    0.300  
#>   y ~                        
#>     m          (b)    0.500  
#>     x                 0.000  
#> 
#> Variances:
#>                    Population
#>    .m                 0.910  
#>    .y                 0.750  
#>     x                 1.000  
#> 
#> Defined Parameters:
#>                    Population
#>     ab                0.150  
#> 
#> ======================= Data Information =======================
#> 
#> Number of Replications:  2 
#> Sample Sizes:  100 
#> 
#> Call print with 'data_long = TRUE' for further information.
#> 
#> ==================== Extra Element(s) Found ====================
#> 
#> - fit
#> 
#> === Element(s) of the First Dataset ===
#> 
#> ============ <fit> ============
#> 
#> lavaan 0.6-19 ended normally after 1 iteration
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                         5
#> 
#>   Number of observations                           100
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> ======== <test_parameters: CIs (standardized) > ========
#> 
#> Mean(s) across replication:
#>   test_label lhs op rhs label    est    se      z pvalue   cilo  cihi   sig
#> 1          a   m  ~   x     a  0.347 0.085  4.069  0.000  0.180 0.514 1.000
#> 2          b   y  ~   m     b  0.446 0.086  5.190  0.000  0.277 0.615 1.000
#> 3        y~x   y  ~   x       -0.024 0.094 -0.233  0.232 -0.209 0.160 0.000
#> 4       m~~m   m ~~   m        0.880 0.059 14.888  0.000  0.764 0.995 1.000
#> 5       y~~y   y ~~   y        0.791 0.071 11.181  0.000  0.652 0.931 1.000
#> 6       x~~x   x ~~   x        1.000 0.000    NaN    NaN  1.000 1.000 1.000
#> 7         ab  ab := a*b    ab  0.155 0.050  3.102  0.002  0.057 0.253 1.000
#> 
#> - The column 'sig' shows the rejection rates.
#> - If the null hypothesis is false, the rate is the power.
#> - Number of valid replications for rejection rate(s): 2 
#> - Proportion of valid replications for rejection rate(s): 1.000 

test_out <- power4test(object = sim_only,
                       test_fun = test_parameters,
                       test_args = list(op = "~"))
#> Displaying progress enabled. Set 'progress = FALSE' to hide the progress.
#> Do the test: test_parameters: CIs (op: ~) 

print(test_out,
      test_long = TRUE)
#> 
#> ====================== Model Information ======================
#> 
#> == Model on Factors/Variables ==
#> 
#> m ~ a*x
#> y ~ b*m + x
#> ab := a * b
#> 
#> == Model on Variables/Indicators ==
#> 
#> m ~ a*x
#> y ~ b*m + x
#> ab := a * b
#> 
#> ====== Population Values ======
#> 
#> Regressions:
#>                    Population
#>   m ~                        
#>     x          (a)    0.300  
#>   y ~                        
#>     m          (b)    0.500  
#>     x                 0.000  
#> 
#> Variances:
#>                    Population
#>    .m                 0.910  
#>    .y                 0.750  
#>     x                 1.000  
#> 
#> Defined Parameters:
#>                    Population
#>     ab                0.150  
#> 
#> ======================= Data Information =======================
#> 
#> Number of Replications:  2 
#> Sample Sizes:  100 
#> 
#> Call print with 'data_long = TRUE' for further information.
#> 
#> ==================== Extra Element(s) Found ====================
#> 
#> - fit
#> 
#> === Element(s) of the First Dataset ===
#> 
#> ============ <fit> ============
#> 
#> lavaan 0.6-19 ended normally after 1 iteration
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                         5
#> 
#>   Number of observations                           100
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> ============ <test_parameters: CIs (op: ~)> ============
#> 
#> Mean(s) across replication:
#>   test_label lhs op rhs label    est    se      z pvalue   cilo  cihi   sig
#> 1          a   m  ~   x     a  0.382 0.103  3.698  0.000  0.179 0.585 1.000
#> 2          b   y  ~   m     b  0.443 0.094  4.715  0.000  0.259 0.627 1.000
#> 3        y~x   y  ~   x       -0.029 0.104 -0.226  0.236 -0.232 0.174 0.000
#> 
#> - The column 'sig' shows the rejection rates.
#> - If the null hypothesis is false, the rate is the power.
#> - Number of valid replications for rejection rate(s): 2 
#> - Proportion of valid replications for rejection rate(s): 1.000